MenuOptions refactored in its own component
