+++ src/common/ui/MenuOptions/index.tsx
@@ -1,0 +1,1 @@
+export { MenuOptions } from "./MenuOptions";
